Refactor filter and Yandex map components for improved layout and functionality
