+++ src/components/yandex-map/components/toggle-items/toggle-sidebar.module.scss
@@ -1,26 +1,6 @@
 @import '../../../../shared/styles/extends.scss';
 @import '../../../../shared/styles/mixins.scss';
 
-.root {
-  position: relative;
-  max-width: 384px;
-  height: 100vh;
-  z-index: 10;
-  max-height: calc(100vh - 155px);
-  z-index: 10;
-  height: 100vh;
-  transform: translateX(-120%);
-  transition: transform 0.3s ease;
-
-  @include media-down(1256px) {
-    max-height: calc(100vh - 74px);
-  }
-
-  @include media-down($mobile-width) {
-    max-width: 80%;
-  }
-}
-
 .items {
   display: flex;
   flex-direction: column;
@@ -34,7 +14,7 @@
   top: 0;
   bottom: 0;
   z-index: 10;
-  overflow-y: scroll;
+  overflow-y: auto;
   overflow-x: hidden;
   transition: transform 0.3s ease-in-out;
   background-color: #fff;
@@ -91,31 +71,4 @@
 
 .price {
   color: var(--primary);
-}
-
-.close {
-  position: absolute;
-  width: 40px;
-  height: 40px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 24px;
-  border: 1px solid var(--primary);
-  color: var(--primary);
-  top: 10px;
-  right: -50px;
-  background-color: #fff;
-  z-index: 10;
-  cursor: pointer;
-
-  svg {
-    fill: var(--primary);
-    width: 24px;
-    height: 24px;
-  }
-}
-
-.open {
-  transform: translateX(0);
 }
