+++ src/components/yandex-map/data.json
@@ -270,5 +270,213 @@
         }
       }
     ]
+  },
+  "filter": {
+    "title": "Фильтр",
+    "items": [
+      {
+        "id": "1",
+        "title": "До метро",
+        "type": "checkbox_btn",
+        "value": [
+          {
+            "id": "1",
+            "value": "Не важно"
+          },
+          {
+            "id": "2",
+            "value": "Пешком"
+          }
+        ]
+      },
+      {
+        "id": "2",
+        "title": "Дополнительно",
+        "type": "checkbox",
+        "value": [
+          {
+            "id": "1",
+            "value": "С детской площадкой"
+          },
+          {
+            "id": "2",
+            "value": "Со спортивной площадкой"
+          }
+        ]
+      },
+      {
+        "id": "3",
+        "title": "Лифты",
+        "type": "checkbox_btn",
+        "value": [
+          {
+            "id": "1",
+            "value": "Не важно"
+          },
+          {
+            "id": "2",
+            "value": "Есть"
+          }
+        ]
+      },
+      {
+        "id": "4",
+        "title": "Мусоропровод",
+        "type": "checkbox_btn",
+        "value": [
+          {
+            "id": "1",
+            "value": "Не важно"
+          },
+          {
+            "id": "2",
+            "value": "Есть"
+          }
+        ]
+      },
+      {
+        "id": "5",
+        "title": "Балкон",
+        "type": "checkbox_btn",
+        "value": [
+          {
+            "id": "1",
+            "value": "Балкон"
+          },
+          {
+            "id": "2",
+            "value": "Лоджия"
+          },
+          {
+            "id": "3",
+            "value": "Нет"
+          }
+        ]
+      },
+      {
+        "id": "6",
+        "title": "Санузел",
+        "type": "checkbox_btn",
+        "value": [
+          {
+            "id": "1",
+            "value": "Раздельный"
+          },
+          {
+            "id": "2",
+            "value": "Совмещенный"
+          }
+        ]
+      },
+      {
+        "id": "7",
+        "title": "Вид из окна",
+        "type": "checkbox_btn",
+        "value": [
+          {
+            "id": "1",
+            "value": "Во двор"
+          },
+          {
+            "id": "2",
+            "value": "На улицу и двор"
+          },
+          {
+            "id": "3",
+            "value": "На улицу"
+          },
+          {
+            "id": "4",
+            "value": "На солнечную сторону"
+          }
+        ]
+      },
+      {
+        "id": "8",
+        "title": "Ремонт",
+        "type": "checkbox_btn",
+        "value": [
+          {
+            "id": "1",
+            "value": "Косметический"
+          },
+          {
+            "id": "2",
+            "value": "Евроремонт"
+          },
+          {
+            "id": "3",
+            "value": "Дизайнерский"
+          },
+          {
+            "id": "4",
+            "value": "Требует ремонта"
+          }
+        ]
+      },
+      {
+        "id": "9",
+        "title": "Высота потолков",
+        "type": "checkbox_btn",
+        "value": [
+          {
+            "id": "1",
+            "value": "от 2,5 м"
+          },
+          {
+            "id": "2",
+            "value": "от 2,7 м"
+          },
+          {
+            "id": "3",
+            "value": "от 3 м"
+          },
+          {
+            "id": "4",
+            "value": "от 4 м"
+          },
+          {
+            "id": "4",
+            "value": "Только высокие"
+          }
+        ]
+      },
+      {
+        "id": "10",
+        "title": "Цена, руб.",
+        "type": "range",
+        "value": [0, 180600000]
+      },
+      {
+        "id": "11",
+        "title": "Площадь, м2",
+        "type": "range",
+        "value": [0, 675]
+      },
+      {
+        "id": "12",
+        "title": "Этаж",
+        "type": "range",
+        "value": [1, 50]
+      },
+      {
+        "id": "13",
+        "title": "Год постройки",
+        "type": "range",
+        "value": [1980, 2024]
+      },
+      {
+        "id": "14",
+        "title": "Метраж жилой площади, м2",
+        "type": "range",
+        "value": [0, 500]
+      },
+      {
+        "id": "15",
+        "title": "Метраж кухни, м2",
+        "type": "range",
+        "value": [0, 500]
+      }
+    ]
   }
 }
