+++ src/components/yandex-map/yandex-map.module.scss
@@ -1,6 +1,28 @@
+@import '../../shared/styles/extends.scss';
+@import '../../shared/styles/mixins.scss';
+
+.aside {
+  position: relative;
+  max-width: 384px;
+  height: 100vh;
+  z-index: 10;
+  max-height: calc(100vh - 155px);
+  z-index: 10;
+  height: 100vh;
+  transform: translateX(-120%);
+  transition: transform 0.3s ease;
+
+  @include media-down(1256px) {
+    max-height: calc(100vh - 74px);
+  }
+
+  @include media-down($mobile-width) {
+    max-width: 80%;
+  }
+}
+
 .root {
   position: relative;
-  overflow: hidden;
 }
 
 .wrapper {
@@ -22,4 +44,35 @@
   height: 40px;
   border-radius: 4px;
   border: 1px solid red;
+}
+
+.filter_hidden {
+  display: none;
+}
+
+.close {
+  position: absolute;
+  width: 40px;
+  height: 40px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 24px;
+  border: 1px solid var(--primary);
+  color: var(--primary);
+  top: 10px;
+  right: -50px;
+  background-color: #fff;
+  z-index: 10;
+  cursor: pointer;
+
+  svg {
+    fill: var(--primary);
+    width: 24px;
+    height: 24px;
+  }
+}
+
+.open {
+  transform: translateX(0);
 }
