+++ src/shared/components/filter/filter.module.scss
@@ -8,8 +8,7 @@
   width: 100%;
   max-width: 370px;
   max-height: calc(100vh - 250px);
-  overflow: hidden;
-  overflow-y: auto;
+
   z-index: 5;
   background-color: var(--white);
   display: flex;
@@ -102,6 +101,11 @@
   }
 }
 
+.content {
+  overflow: hidden;
+  overflow-y: auto;
+}
+
 .view_toggler {
   @extend %flex-align;
   gap: 8px;
@@ -140,9 +144,10 @@
     padding: 3px;
     color: var(--white);
     width: fit-content;
-    min-width: 12px;
-    font-size: 8px;
-    height: 12px;
+    min-width: 20px;
+    font-size: 10px;
+    line-height: 17px;
+    height: 20px;
   }
 }
 
@@ -150,12 +155,8 @@
   display: flex;
   flex-direction: column;
   gap: var(--gap-sm);
-
-  &:not(:last-child) {
-    padding: var(--gap-sm);
-    padding-right: 0;
-    border-bottom: 1px solid rgb(209, 213, 219);
-  }
+  padding: var(--gap-sm);
+  border-bottom: 1px solid rgb(209, 213, 219);
 
   .content {
     display: none;
@@ -181,8 +182,4 @@
   align-items: center;
   justify-content: space-between;
   cursor: pointer;
-
-  svg {
-    width: 40px;
-  }
 }
